requirements.txt
streamlit>=1.28.0
openai>=1.0.0
python-docx
PyPDF2
paddlepaddle>=3.2.0,<3.4.0
paddleocr==3.3.2
mcp>=1.2.0
httpx>=0.27.0
# mcp 在 Windows 上会导入 pywintypes，由 pywin32 提供
pywin32>=306; sys_platform == "win32"
